backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-filter==23.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.26.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2023.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
spectacular==0.4.0
sqlparse==0.4.3
tzdata==2023.2
uritemplate==4.1.1
urllib3==1.26.15
xlwt==1.3.0
